fix: floats not being allowed for slider parameters

+++ settings-parser/src/xml/display_type.rs
@@ -12,9 +12,9 @@ pub enum DisplayType {
     Toggle,
     Options(OptionsArray), // array of displayNames
     Slider {
-        min: i32,
-        max: i32,
-        div: i32
+        min: f32,
+        max: f32,
+        div: f32
     },
     SubtleSeparator
 }
@@ -49,7 +49,7 @@ impl TryFrom<&Node<'_, '_>> for DisplayType {
             } else if spl.len() != 4 {
                 Err(format!("Invalid amount of slider parameters. Should be 3, is {}", spl.len() - 1))
             } else {
-                let min = spl[1].parse::<i32>();
+                let min = spl[1].parse::<f32>();
                 if min.is_err() {
                     return Err(format!("Slider min value parse error: {}", min.unwrap_err()));   
                 }
@@ -57,7 +57,7 @@ impl TryFrom<&Node<'_, '_>> for DisplayType {
                 let min = min.unwrap();
 
 
-                let max = spl[2].parse::<i32>();
+                let max = spl[2].parse::<f32>();
                 if max.is_err() {
                     return Err(format!("Slider max value parse error: {}", max.unwrap_err()));   
                 }
@@ -65,13 +65,13 @@ impl TryFrom<&Node<'_, '_>> for DisplayType {
                 let max = max.unwrap();
 
 
-                let div = spl[3].parse::<i32>();
+                let div = spl[3].parse::<f32>();
                 if div.is_err() {
                     return Err(format!("Slider div value parse error: {}", div.unwrap_err()));   
                 }
 
                 let div = div.unwrap();
-                if div <= 0 {
+                if div <= 0.0 {
                     return Err("Slider div value must be greater than 0".to_string());
                 }
